+++ ximpol/config/crab_pulsar.py
@@ -57,7 +57,7 @@ _phi = 0.5*(_phi0 + _phi1)
 # Build the PL normalization as a function of the phase.
 fmt = dict(xname='Pulsar phase', yname='PL normalization',
            yunits='cm$^{-2}$ s$^{-1}$ keV$^{-1}$')
-pl_normalization = xInterpolatedUnivariateSpline(_phi, _norm, k=3, **fmt)
+pl_normalization_spline = xInterpolatedUnivariateSpline(_phi, _norm, k=3, **fmt)
 
 # Fit the PL index as a function of the phase with a sinusoid.
 def ffit(x, *pars):
@@ -70,10 +70,11 @@ popt, pcov = curve_fit(ffit, _phi, _index, p0, _index_err)
 
 # And use tho best-fit parameters to build a proper spline.
 fmt = dict(xname='Pulsar phase', yname='PL index')
-pl_index = xInterpolatedUnivariateSpline(_phi, ffit(_phi, *popt), k=2, **fmt)
+pl_index_spline = xInterpolatedUnivariateSpline(_phi, ffit(_phi, *popt), k=2,
+                                                **fmt)
 
 # Build the actual energy spectrum.
-energy_spectrum = power_law(pl_normalization, pl_index)
+energy_spectrum = power_law(pl_normalization_spline, pl_index_spline)
 
 # Build the polarization angle as a function of the phase.
 _phi, _pol_angle = numpy.loadtxt(PANG_FILE_PATH, unpack=True)
@@ -81,12 +82,12 @@ _pol_angle = numpy.deg2rad(_pol_angle)
 # Filter the data points to reduce the noise.
 _pol_angle = scipy.signal.wiener(_pol_angle, 7)
 fmt = dict(xname='Pulsar phase', yname='Polarization angle [rad]')
-pol_angle = xInterpolatedUnivariateSpline(_phi, _pol_angle, k=1, **fmt)
+pol_angle_spline = xInterpolatedUnivariateSpline(_phi, _pol_angle, k=1, **fmt)
 
 # Mind that you have to wrap this into a function to be used.
 # This could be done in a more general way with some sort of library function.
 def polarization_degree(E, t, ra, dec):
-    return pol_degree(t)
+    return pol_degree_spline(t)
 
 # Build the polarization degree as a function of the phase.
 _phi, _pol_degree = numpy.loadtxt(PDEG_FILE_PATH, unpack=True)
@@ -94,11 +95,11 @@ _pol_degree /= 100.
 # Filter the data points to reduce the noise.
 _pol_degree = scipy.signal.wiener(_pol_degree, 5)
 fmt = dict(xname='Pulsar phase', yname='Polarization degree')
-pol_degree = xInterpolatedUnivariateSpline(_phi, _pol_degree, k=1, **fmt)
+pol_degree_spline = xInterpolatedUnivariateSpline(_phi, _pol_degree, k=1, **fmt)
 
 # And, again, this needs to be wrapped into a function.
 def polarization_angle(E, t, ra, dec):
-    return pol_angle(t)
+    return pol_angle_spline(t)
 
 
 ROI_MODEL = xROIModel(83.633083, 22.014500)
@@ -113,11 +114,11 @@ if __name__ == '__main__':
     print(ROI_MODEL)
     from ximpol.utils.matplotlib_ import pyplot as plt
     plt.figure()
-    pl_index.plot(show=False)
+    pl_index_spline.plot(show=False)
     plt.figure()
-    pl_normalization.plot(show=False)
+    pl_normalization_spline.plot(show=False)
     plt.figure()
-    pol_angle.plot(show=False)
+    pol_angle_spline.plot(show=False)
     plt.figure()
-    pol_degree.plot(show=False)
+    pol_degree_spline.plot(show=False)
     plt.show()