+++ ximpol/examples/crab_pulsar.py
@@ -26,23 +26,25 @@ from ximpol.core.pipeline import xPipeline
 from ximpol.evt.binning import xBinnedModulationCube, xEventBinningBase
 from ximpol.evt.event import xEventFile
 from ximpol.utils.matplotlib_ import pyplot as plt
-from ximpol.config.crab_pulsar import pol_degree, pol_angle, pl_index,\
-    pl_normalization
+from ximpol.config.crab_pulsar import pol_degree_spline, pol_angle_spline,\
+    pl_index_spline, pl_normalization_spline
 
 
 """Script-wide simulation and analysis settings.
 """
 CFG_FILE_PATH = os.path.join(XIMPOL_CONFIG, 'crab_pulsar.py')
 OUT_FILE_PATH_BASE = os.path.join(XIMPOL_DATA, 'crab_pulsar')
+EVT_FILE_PATH = '%s.fits' % OUT_FILE_PATH_BASE
+ANALYSIS_FILE_PATH = '%s_analysis.txt' % OUT_FILE_PATH_BASE
 SIM_DURATION = 100000.
 NUM_PHASE_BINS = 20
+PHASE_BINNING = None
 E_BINNING = [1., 10.]
 
+
+"""Main pipeline object.
+"""
 PIPELINE = xPipeline(clobber=False)
-
-
-evt_file_path = '%s.fits' % OUT_FILE_PATH_BASE
-analysis_file_path = '%s_analysis.txt' % OUT_FILE_PATH_BASE
 
 
 def _sel_file_path(i):
@@ -64,7 +66,7 @@ def _phase_binning():
     """Read the input event file and create an equipopulated binning in the
     pulsar phase.
     """
-    evt_file = xEventFile(evt_file_path)
+    evt_file = xEventFile(EVT_FILE_PATH)
     phase = evt_file.event_data['PHASE']
     bins = xEventBinningBase.equipopulated_binning(NUM_PHASE_BINS, phase, 0, 1)
     logger.info('Phase binning: %s' % bins)
@@ -75,17 +77,14 @@ def generate():
     """Generate the events.
     """
     PIPELINE.xpobssim(configfile=CFG_FILE_PATH, duration=SIM_DURATION,
-                      outfile=evt_file_path)
-
-# Since we need this in all the following steps, we create the binning at the
-# top level of the module.
-phasebins = _phase_binning()
+                      outfile=EVT_FILE_PATH)
 
 def prepare():
     """Prepare the event data for the actual analysis.
     """
-    for i, (_min, _max) in enumerate(zip(phasebins[:-1], phasebins[1:])):
-        PIPELINE.xpselect(evt_file_path, phasemin=_min, phasemax=_max,
+    for i, (_min, _max) in enumerate(zip(PHASE_BINNING[:-1],
+                                         PHASE_BINNING[1:])):
+        PIPELINE.xpselect(EVT_FILE_PATH, phasemin=_min, phasemax=_max,
                           outfile=_sel_file_path(i))
         PIPELINE.xpbin(_sel_file_path(i), algorithm='MCUBE', ebinalg='LIST',
                        ebinning=E_BINNING, outfile=_mcube_file_path(i))
@@ -95,9 +94,10 @@ def prepare():
 def analyze():
     """Analyze the data.
     """
-    logger.info('Opening output file %s...' % analysis_file_path)
-    analysis_file = open(analysis_file_path, 'w')
-    for i, (_min, _max) in enumerate(zip(phasebins[:-1], phasebins[1:])):
+    logger.info('Opening output file %s...' % ANALYSIS_FILE_PATH)
+    analysis_file = open(ANALYSIS_FILE_PATH, 'w')
+    for i, (_min, _max) in enumerate(zip(PHASE_BINNING[:-1],
+                                         PHASE_BINNING[1:])):
         _mcube = xBinnedModulationCube(_mcube_file_path(i))
         _mcube.fit()
         _fit_results = _mcube.fit_results[0]
@@ -107,13 +107,12 @@ def analyze():
         _pol_deg_err = _fit_results.polarization_degree_error
         _pol_angle = _fit_results.phase
         _pol_angle_err = _fit_results.phase_error
-        _xspec_model = PIPELINE.xpxspec(_pha1_file_path(i))
-        _index = _xspec_model(1).values[0]
-        _index_err = _xspec_model(1).sigma
+        _spec_fitter = PIPELINE.xpxspec(_pha1_file_path(i), plot=False)
+        (_index, _index_err), (_norm, _norm_err) =_spec_fitter.fit_parameters()
         # The division by the phase interval is a workaround and we should
         # keep track of that in xpselect.
-        _norm = _xspec_model(2).values[0]/(_max - _min)
-        _norm_err = _xspec_model(1).sigma/(_max - _min)
+        _norm /= (_max - _min)
+        _norm_err /= (_max - _min)
         _data = (_phase, _phase_err, _pol_deg, _pol_deg_err, _pol_angle,
                  _pol_angle_err, _index, _index_err, _norm, _norm_err)
         _fmt = ('%.4e   ' * len(_data)).strip()
@@ -129,38 +128,40 @@ def plot():
     mod_label = 'Input model'
     _phase, _phase_err, _pol_deg, _pol_deg_err, _pol_angle,\
         _pol_angle_err, _index, _index_err, _norm,\
-        _norm_err = numpy.loadtxt(analysis_file_path, unpack=True)
+        _norm_err = numpy.loadtxt(ANALYSIS_FILE_PATH, unpack=True)
     plt.figure('Polarization degree')
     plt.errorbar(_phase, _pol_deg, xerr=_phase_err, yerr=_pol_deg_err, fmt='o',
                  label=sim_label)
-    pol_degree.plot(show=False, label=mod_label)
+    pol_degree_spline.plot(show=False, label=mod_label)
     plt.axis([None, None, 0., 0.4])
     plt.legend(bbox_to_anchor=(0.45, 0.95))
     plt.figure('Polarization angle')
     plt.errorbar(_phase, _pol_angle, xerr=_phase_err, yerr=_pol_angle_err,
                  fmt='o', label=sim_label)
-    pol_angle.plot(show=False, label=mod_label)
+    pol_angle_spline.plot(show=False, label=mod_label)
     plt.legend(bbox_to_anchor=(0.45, 0.95))
     plt.figure('PL normalization')
     plt.errorbar(_phase, _norm, xerr=_phase_err, yerr=_norm_err, fmt='o',
                  label=sim_label)
-    pl_normalization.plot(show=False, label=mod_label)
+    pl_normalization_spline.plot(show=False, label=mod_label)
     plt.legend(bbox_to_anchor=(0.45, 0.95))
     plt.figure('PL index')
     plt.errorbar(_phase, _index, xerr=_phase_err, yerr=_index_err, fmt='o',
                  label=sim_label)
-    pl_index.plot(show=False, label=mod_label)
+    pl_index_spline.plot(show=False, label=mod_label)
     plt.legend(bbox_to_anchor=(0.45, 0.95))
     plt.show()
 
 def run():
     """Run all the tasks.
     """
-    if os.path.exists(analysis_file_path):
+    if os.path.exists(ANALYSIS_FILE_PATH):
         logger.info('%s exists, delete it if you want to recreate it.' %\
-                    analysis_file_path)
+                    ANALYSIS_FILE_PATH)
     else:
         generate()
+        global PHASE_BINNING
+        PHASE_BINNING = _phase_binning()
         prepare()
         analyze()
     plot()